% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshist.R
\name{rshist}
\alias{rshist}
\title{Random Sample Histogram}
\usage{
rshist(d, n, x, y = NULL, s = NULL, ...)
}
\arguments{
\item{d}{Character variable representing 4 types of distributions (n = Normal, t = t, f = F, and b = Binomial)}

\item{n}{Interger variable for sample size}

\item{x}{Parameter 1 of the distribution}

\item{y}{Parameter 2 of the distribution}

\item{s}{Set seed to s interger}
}
\value{

}
\description{
Allow users to generate plots of histograms for random samples from 4 different distributions: Normal("n"), t("t"), F("f"), Binomial("b")
Users should only have to enter the name of the distribution ("n", "t", "f", "b") and the number of samples (n) to get a result
Users should be able to include additional parameters for each distribution as well as the number of bins for the histogram
Normal: Mean (x) and Standard Deviation (y)
t: Degrees of Freedom (x)
F: Numerator (x) and Denominator Degrees of Freedom (y)
Binomial: Size-Number of Trials (x),and p-Probability of Success on each Trial (y)
Users should be able to choose to set the seed or not and if they set the seed the value of the seed (the last arguments or s).
}
\examples{
rsdist("n", n = 50, x = 6, y = 5, s = 1)
}
